detect if we have a tty in the helpers
When you use the helpers in scripts we get python or other exceptions
from docker-compose or docker itself because at that moment the helpers
no longer have a tty available. Now we only enable the tty on the
docker-compose run or docker run or docker exec when a tty is detected.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>