use execvp instead of subprocess for mysql
If we handoff the execution to the script we are calling instead of
trying to handle it with subprocess we avoid complex situations to pass
thru signals and such.

Also fo allow redirecting into mysql we find the name of the running
mysql container and use that with the docker commandline tool instead of
thru docker-compose which does not handle redirecting into the
application run in the container very well.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>